% Generated by roxygen2 and by hand
\name{xrd_db}
\alias{xrd_db}
\title{Graphical user interface for XRD data entry}
\usage{
xrd_db()
}
\arguments{
\item{xrd_db}{Shine app for XRD databse}
}
\value{
}
\description{
This package provide a graphical user interface to enter XRD data easily into the database.
}
\examples{
xrd_db()
}
\seealso{
}
